microservicio_ingesta/requirements.txt
Django==4.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
djongo==1.3.6
dnspython==2.2.1
requests==2.28.1
pymongo==3.12.3
python-dotenv==0.20.0
mongoengine
pyjwt